# Build: Add load performance check inside perf function (fixes #3994)