**Output:**

```sh
> node Makefile.js perf

CPU Speed is 2793 with multiplier 7500000
Performance Run #1:  1868.417686ms
Performance Run #2:  1840.1341579999998ms
Performance Run #3:  1878.664651ms
Performance Run #4:  1850.484123ms
Performance Run #5:  1843.7707209999999ms
Performance budget ok:  1850.484123ms (limit: 2685.2846401718584ms)


Load performance Run #1:  291.835235ms
Load performance Run #2:  290.304513ms
Load performance Run #3:  298.500252ms
Load performance Run #4:  292.96055ms
Load performance Run #5:  300.357205ms

Load Performance median :  292.96055ms
```